Achieve the next major step with PyHelloWorld example
When everything is included in one dynamic library the bug observed so far is resolved (reason: strings passed over dll boundary). To achieve this, several things have to be adapted in the generated FireBreath configuration as well as in the swig generated module.
A try is done here to automate this by applying patches in a hot monkey way (i.e. on the fly).